it works now

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: labia-fe <labia-fe@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 12:20:50 by labia-fe          #+#    #+#             */
-/*   Updated: 2024/10/29 12:13:46 by labia-fe         ###   ########.fr       */
+/*   Updated: 2024/10/29 14:50:18 by labia-fe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ int	ft_type_detect(va_list parameters, char full_str)
 	else if (full_str == 's')
 		count += ft_putstr(va_arg(parameters, char *));
 	else if (full_str == 'p')
-		count += ft_putptr(va_arg(parameters, unsigned long int));
+		count += ft_putptr(va_arg(parameters, unsigned long long int));
 	else if (full_str == 'i')
 		count += ft_putnbr(va_arg(parameters, int));
 	else if (full_str == 'd')
@@ -40,10 +40,10 @@ int	ft_type_detect(va_list parameters, char full_str)
 	else if (full_str == 'u')
 		count += ft_putnbr_unsigned(va_arg(parameters, unsigned int));
 	else if (full_str == 'x')
-		count += ft_putnbr_hexadecimal(va_arg(parameters, unsigned long int),
+		count += ft_putnbr_hexadecimal(va_arg(parameters, unsigned int),
 				"0123456789abcdef");
 	else if (full_str == 'X')
-		count += ft_putnbr_hexadecimal(va_arg(parameters, unsigned long int),
+		count += ft_putnbr_hexadecimal(va_arg(parameters, unsigned int),
 				"0123456789ABCDEF");
 	else if (full_str == '%')
 		count += ft_putchar(full_str);
