+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: labia-fe <labia-fe@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 17:12:11 by labia-fe          #+#    #+#             */
-/*   Updated: 2024/10/29 10:08:55 by labia-fe         ###   ########.fr       */
+/*   Updated: 2024/10/29 14:49:48 by labia-fe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ int		ft_putchar(char c);
 int		ft_putstr(char *s);
 int		ft_putnbr(int nb);
 int		ft_putnbr_unsigned(unsigned int nb);
-int		ft_putnbr_hexadecimal(unsigned long int nb, char *base);
-int		ft_putptr(unsigned long int nb);
+int		ft_putnbr_hexadecimal(unsigned long long int nb, char *base);
+int		ft_putptr(unsigned long long int nb);
 int		ft_type_detect(va_list parameters, char full_str);
 int		ft_printf(char const *full_str, ...);
