+++ put_things_1.c
@@ -6,7 +6,7 @@
 /*   By: labia-fe <labia-fe@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 15:56:25 by labia-fe          #+#    #+#             */
-/*   Updated: 2024/10/29 11:39:24 by labia-fe         ###   ########.fr       */
+/*   Updated: 2024/10/29 14:40:55 by labia-fe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,8 +37,8 @@ int	ft_putstr(char *s)
 
 int	ft_putnbr(int nb)
 {
-	unsigned int	num;
-	int				count;
+	long int	num;
+	int			count;
 
 	num = nb;
 	count = 0;
